pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = "Perform category classification task"
authors = ["Nitesh Pandey"]

[tool.poetry.dependencies]
python = "3.9.13"
keras = "2.13.1"
tensorflow = "2.13.1"
tensorflow_hub = "0.13.0"
PyYAML = "6.0.2"
pandas = "2.0.3"
scikit-learn = "1.4.2"
transformers = "4.44.2"
jupyter = "1.1.1"
tensorflow-macos = "2.13.1"
tensorflow_model_optimization = "0.8.0"

[tool.poetry.dev-dependencies]
pre-commit = ">=2.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^24.3.0"
mkdocs = ">=1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
